refactor(cli): extract core logic into reusable modules
- Extracted types, utils, and generator logic from dokugen.ts
- Created shared library entry point in index.ts
- Standardized ESM build with .js extensions and declarations
