+++ apps/cli/bin/index.ts
@@ -1,0 +1,4 @@
+export * from "./types.js";
+export * from "./utils.js";
+export * from "./generator.js";
+export { detectProjectType } from "./projectDetect.js";
