+++ apps/cli/tsconfig.json
@@ -11,8 +11,19 @@
     "esModuleInterop": true,
     "forceConsistentCasingInFileNames": true,
     "strict": true,
-    "skipLibCheck": true
+    "skipLibCheck": true,
+    "declaration": true,
+    "declarationMap": true
   },
-  "include": ["bin/**/*", "_tests_/**/*", "scripts/**/*"],
-  "exclude": ["dashboard", "lib", "dist", "node_modules"]
+  "include": [
+    "bin/**/*",
+    "_tests_/**/*",
+    "scripts/**/*"
+  ],
+  "exclude": [
+    "dashboard",
+    "lib",
+    "dist",
+    "node_modules"
+  ]
 }
